% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{bil_interpol_grid_dt}
\alias{bil_interpol_grid_dt}
\title{fast full-grid bilinear interpolation for different values between the same grid}
\usage{
bil_interpol_grid_dt(DT, expand_variable_cols, function_value_cols, DT_new)
}
\description{
Interpolates multiple values in DT onto the grid specified by the coordinates
in DT_new. Standard bilinear interpolation functions are looking up the coordinates
for every value. This is inefficient if you want to interpolate multiple variables
between THE SAME grids, since we need to look up the reference coordinates only once,
which is done by this function. A typical scenario is when you have an ensemble forecast
with K members and D lead times, and map it onto a new grid. The origin grid is assumed to be
regular in the sense that we have lons lon_1,...,lon_n, and lats lat_1,...,lat_m, and the
full grid is given by {(lon_i,lat_j)}_{i in {1,...,n}, j in {1,...,m}}

@param DT data table containing old coordinates lon and lat, the value variables to be interpolated
and expand variables, for each level of which the interpolation should be conducted.
In the example above, the expand variables would be 'ensemble member' and 'date'.
@param expand_variable_cols Character vector containing the column names of the expand variables.
@param function_value_cols Character vector containing the column names of the
variables that should be interpolated. For example c('temperature','precip')
@param DT_new Data table containing the new coordinates

@return a data table
@author Claudio
@export
}
