% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{cube_to_dt}
\alias{cube_to_dt}
\title{Function to convert an array to a data table}
\usage{
cube_to_dt(ar, dims = NULL, val_cn = "value")
}
\arguments{
\item{ar}{array.}

\item{dims}{Either a list of vector-type objects giving the dimensions, or a vector of names of such objects in the parent environment.}

\item{val_cn}{Name of the new column}
}
\description{
Lists an array of any dimension as data table, with one column per dimension and one column with the array values.
A particular feature is that dimensions can be passed as names of vector-type objects in the parent environment, which is frequently useful, see examples.
}
\examples{
date = as.Date('2024-01-01','2020-01-02')
x = 1:3
y = 9:10

array_data = runif(2*3*2) |> array(dim = c(2,3,2))

#call without dimension-variables:
dt1 = cube_to_dt(array_data)
print(dt1)

#call with dimensions given as list:
dimlist = list(date = date,x = x,y = y)
dt2 = cube_to_dt(array_data, dims = dimlist)
print(dt2)

# call with names of vector-objects
dt3 = cube_to_dt(array_data, dims = c('date','x','y'))
print(dt3)

}
