% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{GHA_locs}
\alias{GHA_locs}
\title{Get the Greater Horn of Africa Region considered in CONFER}
\usage{
GHA_locs(
  resolution = "half_deg",
  file_name = "/nr/project/stat/CONFER/Data/ICPAC_region.RData"
)
}
\arguments{
\item{resolution}{Either 'half_deg' or 'full_deg', depending on whether you want all half-degree-gridpoints in the region or only full-degree.}

\item{file_name}{where is the ICPAC_region information stored?}
}
\description{
Get the Greater Horn of Africa Region considered in CONFER
}
\examples{
\dontrun{ #for subsetting a data table dt to the ICPAC locations do this:
setkey(dt,lon,lat)
dt_sub = dt[GHA_locs()]}

}
