% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{global_gha_lat_subset}
\alias{global_gha_lat_subset}
\title{specification of smallest grid fully containing the greater-horn-of-Africa-region, as considered in CONFER}
\usage{
global_gha_lat_subset()
}
\description{
specification of smallest grid fully containing the greater-horn-of-Africa-region, as considered in CONFER
}
