% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{load_dt_monthly}
\alias{load_dt_monthly}
\title{convenient wrapper for loading predictions and observations into a data table}
\usage{
load_dt_monthly(
  var_name = "prec",
  area = NULL,
  system_name = "ERA",
  target_mons = NULL,
  init_mons = NULL,
  load_dir = ifelse(system_name \%in\% c("ERA", "era"), yes =
    "~/bigdisk/SFE/ERA_monthly_compiled/", no =
    "~/bigdisk/CONFER/Systems_monthly_compiled/"),
  resolution = "half_degree"
)
}
\arguments{
\item{var_name}{short name for variable to be loaded}

\item{area}{should the variable be only returned for a certain area, e.g. 'ICPAC' returns it only for the ICPAC region (still, the entire globe is loaded, so it doesn't speed the function up).}

\item{system_name}{For 'era' the ERA-reanalysis data is loaded, otherwise the prediction from the NWP with corresponding name is loaded.}

\item{target_mons}{which months do you want?}

\item{init_mons}{If system_name !- 'era', which initialization months for the predictions do you want to consider?}

\item{load_dir}{path to the files.}

\item{resolution}{'half_degree' or 'full_degree'}
}
\description{
WARNING: This version is outdated and should not be used, use load_cds_monthly_data instead!
}
\details{
#'
}
