% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{loyo}
\alias{loyo}
\title{leave-one-year-out application of any function}
\usage{
loyo(
  dt,
  FUN,
  SDcols,
  bycols = c("lon", "lat"),
  mc.cores = 4,
  newnamesstring = "new",
  args = NULL,
  overwrite = F,
  ...
)
}
\arguments{
\item{dt}{Data table with columns 'year', SDcols and bycols}

\item{FUN}{function to apply}

\item{SDcols}{character vector containing the column names of the columns FUN sould be applied to.}

\item{bycols}{character vector containing the column names of the grouping variables, pass NULL if you don't want to group.}

\item{mc.cores}{how many cores to use in parallelization.}

\item{newnamesstring}{The results of the leave-one-out computation are appended to dt as columns with the names \code{SDcols}_\code{newnamesstring}.}

\item{args}{Character vector. Allows to pass additional columns as arguments to the function. The argument name within FUN needs to match the column name.}

\item{overwrite}{Logical. If you want to use loyo to overwrite existing columns in the data table, this needs to be set to TRUE.}

\item{...}{Arguments passed on to FUN.}
}
\value{
If bycols is not NULL, the data table dt with appended columns, else a data table with the SDcols and a singlerow for each year
}
\description{
new columns are generated in the data table containing the leave-one-out-values. Requires a function with only one argument. If you require multiple arguments use loyoma.
}
