% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{loyo_standardize}
\alias{loyo_standardize}
\title{leave-one-year-out standardization}
\usage{
loyo_standardize(
  dt,
  SDcols,
  bycols = c("lon", "lat"),
  mc.cores = 4,
  newnamesstring = NULL,
  na.rm = T
)
}
\arguments{
\item{dt}{Data table with columns 'year', SDcols and bycols}

\item{SDcols}{character vector containing the column names of the columns that should be standardized.}

\item{bycols}{character vector containing the column names of the grouping variables, pass NULL if you don't want to group.
If you have several ensemble members, do not group by member!}

\item{mc.cores}{how many cores to use in parallelization.}

\item{newnamesstring}{Either NULL or character. If NULL, a data table of the same form as dt is returned with the SDcols
overwritten by standardized versions (default). If character, new columns are attached and their names are
\code{paste0(SDcols,'_',newnamesstring)}.}

\item{na.rm}{Logical. Should NAs be removed when the means and variances are computed?}
}
\value{
a copy of dt with standardized columns.
}
\description{
Takes a data table and standardizes one or several columns in a leave-one-year-out fashion.
For standardization it is important to use leave-one-year-out if one of your columns is 'member', otherwise your mean estimate is in-sample.
}
\author{
Claudio
}
