% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{mask_precip}
\alias{mask_precip}
\title{function for masking out areas with low precip}
\usage{
mask_precip(
  dt,
  threshold = 0.2,
  value_col = "prec",
  bycols = intersect(c("month", "lon", "lat"), names(dt))
)
}
\arguments{
\item{dt}{data table containing precip data}

\item{threshold}{areas with average precip below threshold are masked out}

\item{value_col}{name of the column where precip is stored.}

\item{bycols}{masking is done grouped by these columns (should at least contain 'lon','lat')}
}
\value{
data table dt, with a logical column called 'mask'.
}
\description{
function for masking out areas with low precip
}
\author{
Claudio
}
