% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{permutation_test_difference}
\alias{permutation_test_difference}
\title{function for running a permutation test}
\usage{
permutation_test_difference(a, b, N = 5000)
}
\arguments{
\item{a}{scores of first model}

\item{b}{scores of second model}

\item{N}{numebr of permutations used}
}
\value{
list with three elements: d_bar is the mean score difference, D is a vector with the N mean score differences,
and p_val is the p-value of the permutation test for the hypothesis that a is worse or equally good as b.
}
\description{
function for running a permutation test
}
\author{
Alex
}
