% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{take_ens_mean}
\alias{take_ens_mean}
\title{function for reducing a data table containing ensemble predictions to the ensemble means:}
\usage{
take_ens_mean(dt, pred_vars)
}
\arguments{
\item{dt}{data table, needs to contain a column called 'member'}

\item{pred_vars}{character vector containing the names of the columns containing the ensemble predictions.}
}
\description{
function for reducing a data table containing ensemble predictions to the ensemble means:
}
