% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPutilities.R
\name{time_lag_monthly_dt}
\alias{time_lag_monthly_dt}
\title{function for time-lagging monthly data}
\usage{
time_lag_monthly_dt(dt, lag = 0)
}
\arguments{
\item{dt}{a data table with years and months.}

\item{lag}{how many months do you want to lag?}
}
\value{
The lagged data table.
}
\description{
function for time-lagging monthly data
}
\author{
Claudio
}
